chore: sensible next-auth default w/theme
this migrates the `sendVerificationEmail` to the skill-api package and uses theming capabilities of next-auth to provide basic color and image for email
