+++ packages/create-skill-app/templates/next/package.json
@@ -32,7 +32,6 @@
     "classnames": "^2.3.1",
     "feed": "^4.2.2",
     "groq": "^2.15.0",
-    "mjml": "^4.12.0",
     "next": "^12.2.3",
     "next-auth": "^4.10.2",
     "next-compose-plugins": "^2.2.1",
