+++ packages/skill-api/package.json
@@ -42,6 +42,7 @@
     "date-fns": "^2.28.0",
     "handlebars": "^4.7.7",
     "html-to-text": "^8.2.0",
+    "mjml": "^4.12.0",
     "node-html-markdown": "^1.2.0",
     "preact": "^10.7.2",
     "preact-render-to-string": "^5.2.0",
@@ -62,6 +63,7 @@
     "@types/html-to-text": "^8.1.0",
     "@types/jest": "^27.5.2",
     "@types/lodash": "^4.14.179",
+    "@types/mjml": "^4.7.0",
     "@types/node": "^17.0.0",
     "@types/nodemailer": "^6.4.4",
     "@types/oauth": "^0.9.1",
@@ -79,7 +81,7 @@
     "i": "^0.3.7",
     "msw": "^0.35.0",
     "next": "^12.2.3",
-    "next-auth": "=4.2.1",
+    "next-auth": "^4.10.2",
     "nodemailer": "^6.7.2",
     "nodemailer-postmark-transport": "^5.2.1",
     "postcss": "^8.4.14",
@@ -90,7 +92,7 @@
     "whatwg-fetch": "^3.6.2"
   },
   "peerDependencies": {
-    "next-auth": "=4.2.1",
+    "next-auth": "^4.10.2",
     "nodemailer": "^6.6.5",
     "nodemailer-postmark-transport": "^5.2.1",
     "react": "^17.0.2 || ^18",
