+++ packages/skill-api/process.d.ts
@@ -1,5 +1,6 @@
 declare namespace NodeJS {
   export interface ProcessEnv {
     STRIPE_SECRET_TOKEN: string
+    EMAIL_SERVER_PORT: number
   }
 }
